refactor(api/agents): inline web-search model resolver next to its caller
Cleanup follow-up to the initial #949 commit, no behaviour change.

- Replace lazy `await import()` inside `detectProviderForModelId` with
  top-level imports from `drizzle-orm` + the `aiModels` schema; the
  schema layer is leaf-level so there's no circular-dep risk.
- Move `resolveWebSearchModelId` from
  `agents/subgraphs/research/nodes/shared/` to `agents/core/tools/`
  next to its only consumer (`webSearch.ts`). Removes the backward
  directional dependency from `core` → `subgraphs`.

`bunx tsc --noEmit` clean; 24 research-loop tests + 65 agents tests
still pass.
